requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pandas>=1.3.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
streamlit  # Als je Streamlit gebruikt 
tenacity>=8.0.0
ratelimit>=2.2.1
prometheus-client>=0.12.0  # Voor metrics 